[JTB-19] Count only active users for group statistics

+++ src/main/java/com/github/afanas10101111/jtb/dto/GroupStatTo.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class GroupStatTo {
     private final Integer id;
     private final String title;
-    private final Integer activeUserCount;
+    private final Long activeUserCount;
 }
